requirements.txt
streamlit==1.44.1
PyMuPDF==1.26.0
python-docx==1.1.2
sentence-transformers==3.4.1
faiss-cpu==1.12.0
numpy==2.1.3
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3